KEYCLOAK-16889 Client Policy : Refactor Test Class
